simple form validation

+++ db/migrations/4-create-timer-session.js
@@ -25,7 +25,7 @@ module.exports = {
         },
       },
       time: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.INTEGER(2),
         allowNull: false,
       },
       createdAt: {
